+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 1. add `startIndex` and `endIndex` param
 2. add `skipExistSegments` param
-3. fix: use spawn replace exec
+3. add `suffix` param
+4. fix: use spawn replace exec
 
 # 0.3.1